+++ CHANGELOG.md
@@ -5,13 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-12-17
+## [Unreleased] - 2023-12-25
 ### Changed
 - Migrated getTaskEnablerRoute to TypeScript for improved type safety and maintainability.
 - Migrated scheduler to TypeScript.
 - Modified schedulerTask function to now accept an object as recurrenceRule, enhancing flexibility in scheduling tasks.
 - Migrated schedulerRoute to TypeScript.
 - Migrated scheduledTasksRoute to Typescript
+- Merged config with configs and migrated config to Typescript
 
 ## [v15.5.0] - 2023-12-17
 ### Added